requirements.txt
asgiref==3.7.2
Django==5.0
django-taggit==5.0.1
sqlparse==0.4.4
